+++ source/source_lcao/test/CMakeLists.txt
@@ -19,7 +19,7 @@ AddTest(
     ${ABACUS_SOURCE_DIR}/source_basis/module_ao/parallel_orbitals.cpp
 
     ${ABACUS_SOURCE_DIR}/source_io/module_dm/write_dmr.cpp
-    ${ABACUS_SOURCE_DIR}/source_io/module_output/ucell_io.cpp
+    ${ABACUS_SOURCE_DIR}/source_cell/ucell_io.cpp
     ${ABACUS_SOURCE_DIR}/source_hamilt/module_hcontainer/output_hcontainer.cpp
 )
 
@@ -36,7 +36,7 @@ AddTest(
     ${ABACUS_SOURCE_DIR}/source_hamilt/module_hcontainer/transfer.cpp
     ${ABACUS_SOURCE_DIR}/source_basis/module_ao/parallel_orbitals.cpp
     ${ABACUS_SOURCE_DIR}/source_io/module_dm/write_dmr.cpp
-    ${ABACUS_SOURCE_DIR}/source_io/module_output/ucell_io.cpp
+    ${ABACUS_SOURCE_DIR}/source_cell/ucell_io.cpp
 )
 
 AddTest(
